Celebal_Task_8-main/requirements.txt
django
langchain
pypdf
langchain_groq
langchain_community
langchain_google_genai
faiss-cpu